+++ packages/atproto/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release Note
 
+## v2.4.1
+
+- feat: added `com.atproto.server.describeServer.output.blobUploadLimit`
+- chore: regenerated from synced lexicons
+
 ## v2.4.0
 
 - chore: bump `atproto_core` to `^2.4.0`, picking up the `decodeCar` CAR-header validation, the true 5xx status code on `RetryContext`, the stale-401 refresh short-circuit and `ServiceContext.withAdditionalHeaders`.